add table encode and fix size bugs in base64

+++ base64_src/inc/base64.h
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-bool base64_encode(uint8_t *input, char *output, uint8_t input_size, uint8_t output_size);
-bool base64_decode(char *input, uint8_t *output, uint8_t input_size, uint8_t output_size);
+bool base64_encode(uint8_t *input, char *output, uint16_t input_size, uint16_t output_size);
+bool base64_decode(char *input, uint8_t *output, uint16_t input_size, uint16_t output_size);
 
 #endif
